FileView: moved into components directory
